use 1.10.x process kill logic to send SIGTERM properly to process group. Fix OSS bug that killing finished bash process from BashOperator causes crashes
GitOrigin-RevId: 6fdde6d97be120627e572aa22cc96a7bb7f1a753

+++ airflow/operators/bash_operator.py
@@ -19,6 +19,7 @@
 
 
 import os
+import psutil
 import signal
 from subprocess import Popen, STDOUT, PIPE
 from tempfile import gettempdir, NamedTemporaryFile
@@ -135,4 +136,5 @@ class BashOperator(BaseOperator):
 
     def on_kill(self):
         self.log.info('Sending SIGTERM signal to bash process group')
-        os.killpg(os.getpgid(self.sp.pid), signal.SIGTERM)
+        if self.sp and psutil.pid_exists(self.sp.pid):
+            os.killpg(os.getpgid(self.sp.pid), signal.SIGTERM)
